feat(cutscene-editor): text layer
(cherry picked from commit d88a6b3b7edc6130b983ca0b24518ba22aee2811)

+++ api/src/main/java/fr/loudo/narrativecraft/api/editors/cutscene/layers/CutsceneLayer.java
@@ -119,6 +119,6 @@ public abstract class CutsceneLayer implements ICutsceneLayer {
      */
     protected boolean isExactTick(float tick) {
         if (keyframes.isEmpty()) return false;
-        return keyframes.stream().mapToInt(Keyframe::getTick).max().getAsInt() == tick;
+        return keyframes.stream().mapToInt(Keyframe::getTick).max().getAsInt() == (int) tick;
     }
 }
